chore(sdk-coin-xrp): refactor xrp to own module
- moves xrp/txrp to own module
- imports/re-exports these from bitgo
- moves more types from bitgo into sdk-core

Ticket: BG-48158
