+++ modules/bitgo/package.json
@@ -49,6 +49,7 @@
     "@bitgo/blockapis": "^1.0.1",
     "@bitgo/sdk-api": "^1.0.1",
     "@bitgo/sdk-core": "^1.0.1",
+    "@bitgo/sdk-coin-xrp": "^1.0.0",
     "@bitgo/sjcl": "^1.0.1",
     "@bitgo/statics": "^6.17.0",
     "@bitgo/unspents": "^0.8.0",
@@ -73,10 +74,6 @@
     "lodash": "^4.17.14",
     "moment": "^2.20.1",
     "openpgp": "5.1.0",
-    "ripple-address-codec": "~4.1.3",
-    "ripple-binary-codec": "~0.2.4",
-    "ripple-keypairs": "^0.11.0",
-    "ripple-lib": "~1.4.1",
     "sanitize-html": "^1.27.5",
     "secp256k1": "^4.0.2",
     "secrets.js-grempe": "^1.1.0",
