+++ modules/bitgo/src/config.ts
@@ -19,7 +19,7 @@ import {
   AlgoCoin,
   AvaxERC20Token,
 } from '@bitgo/statics';
-import { EnvironmentName, Environments } from '@bitgo/sdk-core';
+import { EnvironmentName, Environments, KrsProvider } from '@bitgo/sdk-core';
 
 export interface Tokens {
   bitcoin: {
@@ -291,35 +291,6 @@ export const supportedCrossChainRecoveries = {
   bch: ['btc', 'ltc', 'bsv'],
   ltc: ['btc', 'bch', 'bsv'],
   bsv: ['btc', 'ltc', 'bch'],
-};
-
-export type KrsProvider = {
-  feeType: 'flatUsd';
-  feeAmount: number;
-  supportedCoins: string[];
-  feeAddresses?: Record<string, string>
-}
-
-// KRS providers and their fee structures
-export const krsProviders: Record<string, KrsProvider> = {
-  keyternal: {
-    feeType: 'flatUsd',
-    feeAmount: 99,
-    supportedCoins: ['btc', 'eth'],
-    feeAddresses: {
-      btc: '', // TODO [BG-6965] Get address from Keyternal - recovery will fail for now until Keyternal is ready
-    },
-  },
-  bitgoKRSv2: {
-    feeType: 'flatUsd',
-    feeAmount: 0, // we will receive payments off-chain
-    supportedCoins: ['btc', 'eth'],
-  },
-  dai: {
-    feeType: 'flatUsd',
-    feeAmount: 0, // dai will receive payments off-chain
-    supportedCoins: ['btc', 'eth', 'xlm', 'xrp', 'dash', 'zec', 'ltc', 'bch', 'bsv', 'bcha'],
-  },
 };
 
 // TODO: once server starts returning eth address keychains, remove bitgoEthAddress
