+++ modules/bitgo/src/v2/coins/dot.ts
@@ -1,10 +1,10 @@
 import * as accountLib from '@bitgo/account-lib';
 import * as _ from 'lodash';
 import { BitGo } from '../../bitgo';
-import { MethodNotImplementedError } from '../../errors';
 import {
   BaseCoin,
   KeyPair,
+  MethodNotImplementedError,
   ParsedTransaction,
   ParseTransactionOptions,
   SignedTransaction,
