+++ modules/bitgo/src/v2/coins/index.ts
@@ -1,3 +1,5 @@
+import { Xrp, Txrp } from '@bitgo/sdk-coin-xrp';
+export { Xrp, Txrp };
 
 export { AbstractUtxoCoin } from './abstractUtxoCoin';
 export { Algo } from './algo';
@@ -46,10 +48,8 @@ export { Thbar } from './thbar';
 export { Xtz } from './xtz';
 export { Txtz } from './txtz';
 export { Txlm } from './txlm';
-export { Txrp } from './txrp';
 export { Tzec } from './tzec';
 export { Xlm } from './xlm';
-export { Xrp } from './xrp';
 export { Zec } from './zec';
 export { Cspr } from './cspr';
 export { Tcspr } from './tcspr';
