+++ modules/bitgo/src/v2/coins/sol.ts
@@ -9,13 +9,13 @@ import { BaseCoin as StaticsBaseCoin, CoinFamily, coins } from '@bitgo/statics';
 import * as accountLib from '@bitgo/account-lib';
 import { BitGo } from '../../bitgo';
 import * as _ from 'lodash';
-import { MethodNotImplementedError } from '../../errors';
 import { AtaInitializationBuilder } from '@bitgo/account-lib/dist/src/coin/sol';
 import {
   BaseCoin,
   BaseTransaction,
   Memo,
   KeyPair,
+  MethodNotImplementedError,
   ParsedTransaction as BaseParsedTransaction,
   ParseTransactionOptions as BaseParseTransactionOptions,
   SignedTransaction,
