+++ modules/bitgo/src/v2/coins/trx.ts
@@ -11,22 +11,24 @@ import * as request from 'superagent';
 import {
   BaseCoin,
   common,
+  getBip32Keys,
+  getIsKrsRecovery,
+  getIsUnsignedSweep,
   KeyPair,
+  MethodNotImplementedError,
   ParsedTransaction,
   ParseTransactionOptions,
   SignedTransaction,
   SignTransactionOptions,
+  TransactionExplanation,
+  TransactionFee,
+  TransactionPrebuild as BaseTransactionPrebuild,
+  TransactionRecipient as Recipient,
   VerifyAddressOptions,
   VerifyTransactionOptions,
-  TransactionFee,
-  TransactionRecipient as Recipient,
-  TransactionPrebuild as BaseTransactionPrebuild,
-  TransactionExplanation,
 } from '@bitgo/sdk-core';
 
 import { BitGo } from '../../bitgo';
-import { getBip32Keys, getIsKrsRecovery, getIsUnsignedSweep } from '../recovery/initiate';
-import { MethodNotImplementedError } from '../../errors';
 
 export const MINIMUM_TRON_MSIG_TRANSACTION_FEE = 1e6;
