+++ modules/bitgo/src/v2/types.ts
@@ -5,7 +5,3 @@ export interface Recipient {
   address: string;
   amount: string;
 }
-
-export type TokenManagementType = 'enabletoken' | 'disabletoken';
-
-export type ExplainTokenTxType= 'enableToken' | 'disableToken' | 'transferToken';
