+++ modules/sdk-coin-xrp/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@bitgo/sdk-coin-xrp",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "description": "BitGo SDK coin library for Ripple",
   "main": "./dist/src/index.js",
   "types": "./dist/src/index.d.ts",
@@ -38,6 +38,18 @@
     ]
   },
   "dependencies": {
-    "@bitgo/sdk-core": "^1.0.1"
+    "@bitgo/sdk-core": "^1.0.1",
+    "@bitgo/utxo-lib": "^2.2.2",
+    "bignumber.js": "^8.0.1",
+    "bip32": "^2.0.6",
+    "lodash": "^4.17.14",
+    "ripple-address-codec": "~4.1.3",
+    "ripple-binary-codec": "~0.2.4",
+    "ripple-keypairs": "^0.11.0",
+    "ripple-lib": "~1.4.1",
+    "url": "^0.11.0"
+  },
+  "devDependencies": {
+    "@types/lodash": "^4.14.151"
   }
 }
