+++ modules/sdk-coin-xrp/src/index.ts
@@ -1,0 +1,2 @@
+export * from './xrp';
+export * from './txrp';
