+++ modules/sdk-coin-xrp/tsconfig.json
@@ -12,6 +12,9 @@
   "references": [
     {
       "path": "../sdk-core"
+    },
+    {
+      "path": "../utxo-lib"
     }
   ]
 }
