+++ modules/sdk-core/src/bitgo/index.ts
@@ -5,6 +5,7 @@ export * from './baseCoin';
 export * from './bip32util';
 export * from './bitcoin';
 export * from './bitgoBase';
+export * from './config';
 export * from './ecdh';
 export * from './enterprise';
 export * from './environments';
