+++ modules/sdk-core/src/bitgo/recovery/index.ts
@@ -1,7 +1,1 @@
-export interface InitiateRecoveryOptions {
-  userKey: string;
-  backupKey: string;
-  bitgoKey?: string; // optional for xrp recoveries
-  recoveryDestination: string;
-  walletPassphrase?: string;
-}
+export * from './initiate';
