+++ modules/sdk-core/src/bitgo/types.ts
@@ -1,1 +1,5 @@
 export type V1Network = 'bitcoin' | 'testnet';
+
+export type TokenManagementType = 'enabletoken' | 'disabletoken';
+
+export type ExplainTokenTxType = 'enableToken' | 'disableToken' | 'transferToken';
