+++ modules/sdk-core/src/bitgojsError.ts
