+++ modules/sdk-core/src/index.ts
@@ -1,6 +1,6 @@
 export * from './account-lib';
 export * from './api';
 export * from './bitgo';
-export { BitGoJsError } from './bigojsError';
+export * from './bitgojsError';
 import * as common from './common';
 export { common };
