+++ modules/sdk-core/tsconfig.json
@@ -12,6 +12,9 @@
   "references": [
     {
       "path": "../statics"
+    },
+    {
+      "path": "../utxo-lib"
     }
   ]
 }
